Fix: TextClip size 파라미터도 정수 변환
## 문제
- 여전히 'float' object cannot be interpreted as an integer 발생
- size 파라미터의 너비 값도 float였음

## 수정
size 파라미터의 너비를 정수로 변환:
```python
size=(int(self.config.resolution[0] * 0.9), None)
```

## 정수 변환이 필요한 모든 값
✅ size 너비: int(resolution[0] * 0.9)
✅ y_position: int(resolution[1] - 100)

+++ core/editor.py
@@ -356,7 +356,7 @@ class VideoEditor:
                     stroke_color='black',
                     stroke_width=2,
                     method='caption',
-                    size=(self.config.resolution[0] * 0.9, None)
+                    size=(int(self.config.resolution[0] * 0.9), None)
                 )
 
                 # 위치 설정 (하단 중앙, margin 대신 수동 계산)
